docstring updates
