+++ openpathsampling/tests/testopenmmengine.py
@@ -11,8 +11,8 @@ from test_helpers import (true_func, data_filename,
                           assert_not_equal_array_array)
 from openpathsampling.openmm_engine import *
 from openpathsampling.snapshot import Snapshot
-from openpathsampling.snapshot import Momentum, Configuration
-
+from openpathsampling.snapshot import Momentum
+from openpathsampling import Configuration, Momentum
 import simtk.openmm as mm
 from simtk.openmm import app
 from simtk import unit
